+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Level of component changed from Int to ComponentLevel
 - Amount of components changed from Int to ComponentAmount
+- Component name changed from String to ComponentName
+- Component description changed from String to ComponentDescription
 
 ## [0.1.0] - 2019-05-11
 ### Added